Move all cas1 controllers into `cas1.controller`
